feat(content): divide detail
