+++ src/admin/content/views/ContentDetail.tsx
@@ -1,49 +1,46 @@
 import { useMutation } from '@apollo/react-hooks';
-import { compact, flatten, get } from 'lodash-es';
-import React, { FunctionComponent, ReactElement, useCallback, useEffect, useState } from 'react';
-import { Trans, useTranslation } from 'react-i18next';
+import { get } from 'lodash-es';
+import React, {
+	FunctionComponent,
+	ReactElement,
+	ReactText,
+	useCallback,
+	useEffect,
+	useState,
+} from 'react';
+import { useTranslation } from 'react-i18next';
 import MetaTags from 'react-meta-tags';
 
 import {
-	BlockHeading,
+	Blankslate,
 	Button,
 	ButtonToolbar,
 	Container,
+	DropdownButton,
+	DropdownContent,
 	LinkTarget,
+	MenuContent,
 	Navbar,
-	Spacer,
-	Table,
 	Tabs,
-	TagInfo,
-	TagList,
-	TagOption,
-	Thumbnail,
 } from '@viaa/avo2-components';
-import { Avo } from '@viaa/avo2-types';
 
 import { DefaultSecureRouteProps } from '../../../authentication/components/SecuredRoute';
 import { GENERATE_SITE_TITLE } from '../../../constants';
 import { ContentPage } from '../../../content-page/views';
 import {
+	ControlledDropdown,
 	DeleteObjectModal,
 	LoadingErrorLoadedComponent,
 	LoadingInfo,
 } from '../../../shared/components';
 import {
+	createDropdownMenuItem,
 	CustomError,
 	navigate,
 	navigateToAbsoluteOrRelativeUrl,
-	sanitize,
-	sanitizePresets,
 } from '../../../shared/helpers';
 import { useTabs } from '../../../shared/hooks';
 import { ApolloCacheManager, ToastService } from '../../../shared/services';
-import { fetchAllUserGroups } from '../../../shared/services/user-groups-service';
-import {
-	renderDateDetailRows,
-	renderDetailRow,
-	renderSimpleDetailRows,
-} from '../../shared/helpers/render-detail-fields';
 import {
 	AdminLayout,
 	AdminLayoutBody,
@@ -51,12 +48,13 @@ import {
 	AdminLayoutTopBarRight,
 } from '../../shared/layouts';
 import ShareContentPageModal from '../components/ShareContentPageModal';
-import { CONTENT_PATH, GET_CONTENT_DETAIL_TABS, GET_CONTENT_WIDTH_OPTIONS } from '../content.const';
+import { CONTENT_PATH, GET_CONTENT_DETAIL_TABS } from '../content.const';
 import { DELETE_CONTENT } from '../content.gql';
 import { ContentService } from '../content.service';
 import { ContentDetailParams, DbContent } from '../content.types';
 
 import './ContentDetail.scss';
+import { ContentDetailMetaData } from './ContentDetailMetaData';
 
 interface ContentDetailProps extends DefaultSecureRouteProps<ContentDetailParams> {}
 
@@ -64,14 +62,15 @@ const ContentDetail: FunctionComponent<ContentDetailProps> = ({ history, match, 
 	const { id } = match.params;
 
 	// Hooks
+	const [t] = useTranslation();
+
 	const [contentPage, setContentPage] = useState<DbContent | null>(null);
 	const [loadingInfo, setLoadingInfo] = useState<LoadingInfo>({ state: 'loading' });
 	const [isConfirmModalOpen, setIsConfirmModalOpen] = useState<boolean>(false);
-	const [allUserGroups, setAllUserGroups] = useState<TagInfo[]>([]);
 	const [isShareModalOpen, setIsShareModalOpen] = useState<boolean>(false);
+	const [isOptionsMenuOpen, setIsOptionsMenuOpen] = useState<boolean>(false);
 
 	const [triggerContentDelete] = useMutation(DELETE_CONTENT);
-	const [t] = useTranslation();
 
 	const [currentTab, setCurrentTab, tabs] = useTabs(
 		GET_CONTENT_DETAIL_TABS(),
@@ -115,68 +114,6 @@ const ContentDetail: FunctionComponent<ContentDetailProps> = ({ history, match, 
 			});
 		}
 	}, [contentPage, setLoadingInfo]);
-
-	// Get labels of the contentPages, so we can show a readable error message
-	useEffect(() => {
-		fetchAllUserGroups()
-			.then(userGroups => {
-				setAllUserGroups(userGroups);
-			})
-			.catch((err: any) => {
-				console.error('Failed to get user groups', err);
-				ToastService.danger(
-					t(
-						'admin/shared/components/user-group-select/user-group-select___het-controleren-van-je-account-rechten-is-mislukt'
-					),
-					false
-				);
-			});
-	}, [setAllUserGroups, t]);
-
-	// Methods
-	const getUserGroups = (contentPage: Avo.Content.Content): TagOption[] => {
-		const tagInfos: TagInfo[] = compact(
-			(contentPage.user_group_ids || []).map((userGroupId: number): TagInfo | undefined => {
-				return allUserGroups.find(userGroupOption => userGroupOption.value === userGroupId);
-			})
-		);
-		const tagOptions = tagInfos.map(
-			(ug: TagInfo): TagOption => ({
-				id: ug.value,
-				label: ug.label,
-			})
-		);
-		if (tagOptions && tagOptions.length) {
-			return tagOptions;
-		}
-		return [
-			{
-				id: -3,
-				label: t('admin/menu/components/menu-edit-form/menu-edit-form___niemand'),
-			},
-		];
-	};
-
-	const getLabels = (contentPage: DbContent): TagOption[] => {
-		return flatten(
-			(contentPage.content_content_labels || []).map(
-				(contentLabelLink: Avo.Content.ContentLabelLink) => {
-					return contentLabelLink.content_label;
-				}
-			)
-		);
-	};
-
-	const getContentPageWidthLabel = (contentPage: Avo.Content.Content): string => {
-		return (
-			get(
-				GET_CONTENT_WIDTH_OPTIONS().find(
-					option => option.value === contentPage.content_width
-				),
-				'label'
-			) || '-'
-		);
-	};
 
 	const handleDelete = () => {
 		triggerContentDelete({
@@ -235,6 +172,7 @@ const ContentDetail: FunctionComponent<ContentDetailProps> = ({ history, match, 
 				newContentPage,
 				contentPage,
 			});
+
 			ToastService.danger(
 				t(
 					'admin/content/views/content-detail___het-opslaan-van-de-publiek-status-van-de-content-pagina-is-mislukt'
@@ -245,179 +183,108 @@ const ContentDetail: FunctionComponent<ContentDetailProps> = ({ history, match, 
 		setIsShareModalOpen(false);
 	};
 
-	// Render
-	const renderContentDetail = (): ReactElement | null => {
-		if (!contentPage) {
-			return null;
-		}
-		// TODO: Move tab contents to separate views
-		switch (currentTab) {
-			case 'inhoud':
-				return <ContentPage contentPage={contentPage} />;
-			case 'metadata':
-				return (
-					<Container mode="vertical" size="small">
-						<Container mode="horizontal">
-							{!!contentPage.description && (
-								<Spacer margin="bottom-large">
-									<BlockHeading type="h4">
-										<Trans i18nKey="admin/content/views/content-detail___omschrijving">
-											Omschrijving:
-										</Trans>
-									</BlockHeading>
-									<p
-										dangerouslySetInnerHTML={{
-											__html: sanitize(
-												contentPage.description,
-												sanitizePresets.link
-											),
-										}}
-									/>
-								</Spacer>
-							)}
+	const CONTENT_DROPDOWN_ITEMS = [
+		createDropdownMenuItem(
+			'duplicate',
+			t('collection/views/collection-detail___dupliceer'),
+			'copy'
+		),
+		...(!isContentProtected || (isContentProtected && isAdminUser)
+			? [
+					createDropdownMenuItem(
+						'delete',
+						t('admin/content/views/content-detail___verwijderen')
+					),
+			  ]
+			: []),
+	];
 
-							<Table horizontal variant="invisible" className="c-table_detail-page">
-								<tbody>
-									{/*TODO remove cast after typings update to 2.15.0*/}
-									{renderDetailRow(
-										<div style={{ width: '400px' }}>
-											<Thumbnail
-												category="item"
-												src={(contentPage as any).thumbnail_path}
-											/>
-										</div>,
-										t('admin/content/views/content-detail___cover-afbeelding')
-									)}
-									{renderSimpleDetailRows(contentPage, [
-										['title', t('admin/content/views/content-detail___titel')],
-										[
-											'description',
-											t('admin/content/views/content-detail___beschrijving'),
-										],
-										[
-											'content_type',
-											t('admin/content/views/content-detail___content-type'),
-										],
-										['path', t('admin/content/views/content-detail___pad')],
-										[
-											'is_protected',
-											t(
-												'admin/content/views/content-detail___beschermde-pagina'
-											),
-										],
-									])}
-									{renderDetailRow(
-										getContentPageWidthLabel(contentPage),
-										t('admin/content/views/content-detail___breedte')
-									)}
-									{renderDetailRow(
-										`${get(contentPage, 'profile.user.first_name')} ${get(
-											contentPage,
-											'profile.user.last_name'
-										)}`,
-										t('admin/content/views/content-detail___auteur')
-									)}
-									{renderDetailRow(
-										get(contentPage, 'profile.user.role.label'),
-										t('admin/content/views/content-detail___auteur-rol')
-									)}
-									{renderDateDetailRows(contentPage, [
-										[
-											'created_at',
-											t('admin/content/views/content-detail___aangemaakt'),
-										],
-										[
-											'updated_at',
-											t(
-												'admin/content/views/content-detail___laatst-bewerkt'
-											),
-										],
-										[
-											'publish_at',
-											t('admin/content/views/content-detail___gepubliceerd'),
-										],
-										[
-											'depublish_at',
-											t(
-												'admin/content/views/content-detail___gedepubliceerd'
-											),
-										],
-									])}
-									{renderDetailRow(
-										<TagList
-											swatches={false}
-											selectable={false}
-											closable={false}
-											tags={getUserGroups(contentPage)}
-										/>,
-										t('admin/content/views/content-detail___toegankelijk-voor')
-									)}
-									{renderDetailRow(
-										<TagList
-											swatches={false}
-											selectable={false}
-											closable={false}
-											tags={getLabels(contentPage)}
-										/>,
-										t('admin/content/views/content-detail___labels')
-									)}
-								</tbody>
-							</Table>
-						</Container>
-					</Container>
-				);
+	const executeAction = async (item: ReactText) => {
+		switch (item) {
+			case 'duplicate':
+				// TODO: Duplicate content
+				break;
+
+			case 'delete':
+				setIsConfirmModalOpen(true);
+				break;
 
 			default:
 				return null;
 		}
 	};
 
+	const renderContentActions = () => (
+		<ButtonToolbar>
+			<Button
+				type="secondary"
+				icon={get(contentPage, 'is_public') === true ? 'unlock-3' : 'lock'}
+				label={t('admin/content/views/content-detail___publiceren')}
+				title={t(
+					'admin/content/views/content-detail___maak-de-content-pagina-publiek-niet-publiek'
+				)}
+				ariaLabel={t(
+					'admin/content/views/content-detail___maak-de-content-pagina-publiek-niet-publiek'
+				)}
+				onClick={() => setIsShareModalOpen(true)}
+			/>
+			<Button
+				type="secondary"
+				icon="eye"
+				label={t('admin/content/views/content-detail___preview')}
+				title={t('admin/content/views/content-detail___bekijk-deze-pagina-in-de-website')}
+				ariaLabel={t(
+					'admin/content/views/content-detail___bekijk-deze-pagina-in-de-website'
+				)}
+				onClick={handlePreviewClicked}
+			/>
+			<Button
+				label={t('admin/content/views/content-detail___bewerken')}
+				title={t('admin/content/views/content-detail___bewerk-deze-content-pagina')}
+				onClick={() => navigate(history, CONTENT_PATH.CONTENT_EDIT, { id })}
+			/>
+			<ControlledDropdown
+				isOpen={isOptionsMenuOpen}
+				menuWidth="fit-content"
+				onOpen={() => setIsOptionsMenuOpen(true)}
+				onClose={() => setIsOptionsMenuOpen(false)}
+				placement="bottom-end"
+			>
+				<DropdownButton>
+					<Button
+						type="secondary"
+						icon="more-horizontal"
+						ariaLabel={t('collection/views/collection-detail___meer-opties')}
+						title={t('collection/views/collection-detail___meer-opties')}
+					/>
+				</DropdownButton>
+				<DropdownContent>
+					<MenuContent menuItems={CONTENT_DROPDOWN_ITEMS} onClick={executeAction} />
+				</DropdownContent>
+			</ControlledDropdown>
+		</ButtonToolbar>
+	);
+
+	// Render
+	const renderContentDetail = (contentPage: DbContent | null): ReactElement | null => {
+		if (!contentPage) {
+			ToastService.danger(t('De content pagina kon niet worden ingeladen.'), false);
+			return null;
+		}
+
+		switch (currentTab) {
+			case 'inhoud':
+				return <ContentPage contentPage={contentPage} />;
+			case 'metadata':
+				return <ContentDetailMetaData contentPage={contentPage} />;
+			default:
+				return <Blankslate title={t('Dit tabblad kon niet gevonden worden.')} />;
+		}
+	};
+
 	return (
 		<AdminLayout showBackButton pageTitle={pageTitle}>
-			<AdminLayoutTopBarRight>
-				<ButtonToolbar>
-					<Button
-						type="secondary"
-						icon={get(contentPage, 'is_public') === true ? 'unlock-3' : 'lock'}
-						label={t('admin/content/views/content-detail___publiceren')}
-						title={t(
-							'admin/content/views/content-detail___maak-de-content-pagina-publiek-niet-publiek'
-						)}
-						ariaLabel={t(
-							'admin/content/views/content-detail___maak-de-content-pagina-publiek-niet-publiek'
-						)}
-						onClick={() => setIsShareModalOpen(true)}
-					/>
-					<Button
-						type="secondary"
-						icon="eye"
-						label={t('admin/content/views/content-detail___preview')}
-						title={t(
-							'admin/content/views/content-detail___bekijk-deze-pagina-in-de-website'
-						)}
-						ariaLabel={t(
-							'admin/content/views/content-detail___bekijk-deze-pagina-in-de-website'
-						)}
-						onClick={handlePreviewClicked}
-					/>
-					<Button
-						label={t('admin/content/views/content-detail___bewerken')}
-						title={t('admin/content/views/content-detail___bewerk-deze-content-pagina')}
-						onClick={() => navigate(history, CONTENT_PATH.CONTENT_EDIT, { id })}
-					/>
-					{/* TODO: also check permissions */}
-					{(!isContentProtected || (isContentProtected && isAdminUser)) && (
-						<Button
-							label={t('admin/content/views/content-detail___verwijderen')}
-							title={t(
-								'admin/content/views/content-detail___verwijder-deze-content-pagina'
-							)}
-							onClick={() => setIsConfirmModalOpen(true)}
-							type="danger-hover"
-						/>
-					)}
-				</ButtonToolbar>
-			</AdminLayoutTopBarRight>
+			<AdminLayoutTopBarRight>{renderContentActions()}</AdminLayoutTopBarRight>
 			<AdminLayoutHeader>
 				<Navbar background="alt" placement="top" autoHeight>
 					<Container mode="horizontal">
@@ -439,7 +306,7 @@ const ContentDetail: FunctionComponent<ContentDetailProps> = ({ history, match, 
 					<LoadingErrorLoadedComponent
 						loadingInfo={loadingInfo}
 						dataObject={contentPage}
-						render={renderContentDetail}
+						render={() => renderContentDetail(contentPage)}
 					/>
 					<DeleteObjectModal
 						deleteObjectCallback={handleDelete}
